+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 ## [0.1.2] - TBD
 - Added LC type.
-- Adding `foldLoading` extension method to `LCE`, `UCE` and `UCT` types. 
+- Added `foldLoading` extension method to `LCE`, `UCE` and `UCT` types. 
+- Added `foldError` extension method to `LCE`, `UCE` and `UCT` types.
+
 
 ## [0.1.1] - January 5, 2021
 - Added UCE type.